requirements.txt
certifi==2018.8.13
chardet==3.0.4
idna==2.7
mysql-connector==2.1.6
numpy==1.15.0
pytz==2018.5
requests==2.19.1
urllib3==1.23
